Gestion du filtrage dans les sélects au niveau des formulaire d'ajout d'élément
Au niveau par exemple du formulaire d'ajout d'une action par exemple  on doit sélectionner un objectif spécifique, le module de filtrage s'assure donc d'ajouter un select pour le choix de sous programme, choix qui servira à filtrer les objectifs spécifique dans le select des objectifs spécifique.

Ce module est géré de façon à tenir compte des clés étrangères et à pouvoir s'adapter pour toutes les différentes pages, sans une autre intervention spécialisé.

+++ core/Model.php
@@ -110,6 +110,22 @@ class Model {
     }
 
     /**
+     * Reciproque de la function précédente
+     * @return $data array Le tableau de données de base qui avait été assigné au niveau de la function hydrate
+     */
+    public function reciprocHydrate() {
+        
+        $data = [];
+        $data["id"] = $this->getId();
+        foreach ($this->fillable as $attribute) {
+            $methode = 'get' . ucfirst($attribute);
+            $data[$attribute] = $this->$methode();
+        }
+        
+        return $data;
+    }
+
+    /**
      * Méthode permettant de savoir si l'élément du modèle est nouveau ou non.
      * @return bool
      */
@@ -134,6 +150,7 @@ class Model {
         $connection = new DBConn("$tableName");
         if($id = $connection->create($instance->attributes)) {
             $instance->hydrate(["id" => $id]);
+            
             return $instance;
         }
         return null;
@@ -243,14 +260,16 @@ class Model {
         $id = $this->id() ?? null;
 
         if ($id !== null) {
-            if($this->connection->delete($id)) {
+
+            if($this->referenced())
+                return null;
+
+            if($this->connection->delete($id))
                 return $this;
-            }
-            return null;
+        
         }
 
-        throw new Exception("Vous ne pouvez pas supprimer un élément qui n'existe pas !", 1);
-        return false;
+        return null;
     }
 
 }
